fix, tests: succeeding symmetry tests
It looks like the new edits introduced in slabs up to 42d25231c83afb896705fa0f5d08a1cef8051b92 fixed tests of symmetry detection that concern a doubly-thick Fe3O4 slab.

To be honest, I'm not sure what fixed the tests exactly. It may either be the correct use of atom numbers now, or the correct way in which layers are ported over when adding bulk units.