reafctor, fix: getCandidateLayerPeriod
- snake_case_name (also, plural, since we return multiple periodicities)
- __doc__, comments, small clean up

The primary change is to consider a candidate period inappropriate if the two sublayers being tested have different elements or different number of atoms. This was already done in the previous implementation when finding the first bunch of candidates (i.e. comparing sublayers[0] with sublayers[N]), but not when filtering the candidates by looking at all the sublayers in between (i.e., comparing sublayers[i] and sublayers[i+N]))

@fkraushofer: I still have one doubt about whether we should also stop at half the number of layers when we create the first candidates [lines 162--170]. Right now we only stop based on unit-cell height. I could not come up with an example in which stopping at ceil(n_sublayers) / 2 would cause problems. What do you think?