feat: (redo) BulkSlab.apply_bulk_cell_reduction
This method is a reworked extract of parts of SurfaceSlab.makeBulkSlab, but such that it is also compatible with other occurrences of similar repeated code. For now unused, but will be soon.

Algorithmic change: remove duplicates **before** re-centering the slab, as this reduces the number of atoms to iterate over. I don't think this should have any impact on the logics.