refactor, feat, fix: SurfaceSlab.detect_bulk
- clean up code, snake_case_name, __doc__
- use a bunch of the methods developed recently to clarify code
- split off the part that calculates the cuts to improve readability

Functional changes and fixes:
- raise if BULK_LIKE_BELOW is not available
- when making the 'fake bulk', do not recenter its atom coordinates. This is important in order to get the right bulk repeat!
- changed Returns (see __doc__)

Most important change:
The rpars passed in is **modified** with the new SUPERLATTICE found (as part of ensure_minimal_bulk_ab_cell->_change_bulk_cell) and the new BULK_REPEAT.

The modification of rpars.SUPERLATTICE was already done in the version in master.
The new version also changes BULK_REPEAT.

I would very much like the opinion of @fkraushofer and @amimre on this, especially considering my comment next to the function signature (see also the docstring).

Two options:
1. we don't modify anything of rpars, and callers have to do all the legwork (SUPERLATTICE, BULK_REPEAT, LAYER_CUTS, N_BULK_LAYERS) + createLayers + make_bulk_slab. This seems like a lot of work, and potentially repeated code.
2. Do all of the above ourselves in here. This would have the advantage that a successful detect_bulk leaves the slab and rpars ready to be used. To my feeling, a user would naturally expect that calling detect_bulk leaves the slab with the correct .bulkslab available.