feat: ParametersReader
File reader class that will soon help to remove code duplication in parameters._read.

One feature on which I'd like the opinion of  @amimre and @fkraushofer:
We had in the TODOs of parameters for a while to add a complaint about lines that contain a valid PARAMETER but no = sign. (Personally, I have often forgotten one.) I have just implemented this in ParameterReader._warn_about_missing_equals(). The current behavior is to LOG a WARNING. the question is: is this sufficient/too much? Should it be logging.error? Or even an exception?