fix, feat: SurfaceSlab.detect_bulk
As discussed in 9ab4c1136cb499e9d3a04a41c8a94e0a6b69e0d4, now detect_bulk takes care of updating all the information in self and in rpars that come from the detection of the bulk.

A major difference: do not return BULK_REPEAT any more, as it's stored in the given rpars if the detection is successful.

FIXES:
- use the copy of rpars until we're sure to have been successful

Notice that we still have to work with a copy of the slab, as we may error out early with a NoBulkRepeatError and we don't want to exit with a messed-up slab.

@fkraushofer: I have just realized that we do not write SUPERLATTICE out to file (neither when detecting bulk nor later in init). Is this intended? (because we anyway check every time we run)