feat: parameters.comment_out
New function. Nicer to be explicit rather than using modify without a 'new'.

One consequence is that the default of the 'new' argument for parameters._write.modify is now changed from an empty string to a module global (a weird one). This will soon also allow me to change modify to be 'smarter'.