fix: SYMMETRY_EPS (+ cleanup)
When a single value, we were NOT using it for both directions. This means that _Z was always fixed at its default (0.1).

Modified the tests to show the problem.

The current solution is not ideal, though. If we ever change SYMMETRY_EPS and not update _Z, the two would again be different.

A simple solution would be to have a private attribute for Rparams that is assigned to, and a property that returns the _EPS if, e.g., the private attribute is None. Another nice solution would be to have a very simple class for SYMMETRY_EPS that has a .z attribute returning the the non-z if z is None.