refator: rparams module
Split apart the contents of the rparams module into a few bits:

Moved SearchPar to its own module (classes.searchpar.py). Did not make much sense to me to have this as part of rparams, but I'm open for discussion

Split off the other bits of rparams.py into their own submodules of rparams. This helps to confine stuff to its own logical bit.

Created the special subpackage, where not-so-simple parameters can be defined as dedicated classes