fix, tests: turn approx around
pytest.approx uses the type of the value it is passed to infer how things should be checked. Passing a sequence value shortcuts the check with __len__. This is a problem if the other value is an iterable but not a sequence (i.e., it does not have __len__).

The solution is to use approx on the object that may be of non-standard type