feat: Rparams.update
Slightly improved version of the preset loading that was done in tleedm. Allows loading parameter values into rparams, but does not rely on feeding it the EXACT format that the attributes have. For now, it is still a bit restricted for non-special parameters, but could be expanded easily to use a ParametersInterpreter. Perhaps it then would make sense to have a ParametersPresetLoader that handles the interface, deciding if stuff needs to be interpreted from string or is OK as is.