feat: EnergyRange.from_sorted_grid
Make an EnergyRange given a sorted sequence of equally-spaced values (no checks about equality of spacing, though).

This method is useful for the many repeated instances of code in iosearch (writeRfInfo), iorfactor (writeWEXPEL, writeRfactPARAM), and rfactor (run_new_rfactor) that make 'ranges' out of energy grids.

The current implementation for the step comes from those bits of code. However, I'm wondering whether taking the difference between two neighboring energies is the best way to do this. I'd like the opinion of @fkraushofer and @amimre on the comment in from_sorted_grid.