feat: IVShiftRange SpecialParameter
A simple subclass of EnergyRange. Also, replaced everywhere else. I have also done some clean up of a few sections in iorfactor, iosearch, and rfactor that used pretty much the same code. In doing so, I have realized a few things that I'd like to discuss with @fkraushofer and @amimre. Many of the points are found as comments in the code.

My primary problem is that I'm not sure I understand the portion of code in  writeWEXPEL/writeRfInfo/run_new_rfactor that prepares minen and maxen.

If I interpreted the FORTRAN code correctly, the values are just bounds for reading the theory beams. I don't understand why we take a sort of 'intersection' of ranges rather than a 'union'. (Also, I'm not sure the 'intersection' is done in the right way.)

Notice that the version of the code in writeRfactPARAM takes a sort of 'union' (although it does not consider IV_SHIFT).