fix, feat: IVShiftRange bounds step multiples
According to the FORTRAN code, the bounds of the IV_SHIFT_RANGE must be integer multiples of the interpolation step. See especially the divide-then-truncate-to-int in subroutine COMNEI of rfacsb.f: there, the experimental (EE) and theory energies (ET) are already the interpolated ones. If the shift (V0) is not a multiple of EINCR, the indices turn out wrong.