feat, fix: always use rp.THEO_ENERGIES
As discussed in 9d784aa588c97cd23e038da8faee29b5c64ba46a and in person, it is cleaner to always use the THEO_ENERGIES parameter as the range over which R-factor calculations are performed.

There is only one use case where this may differ from the energies in theobeams (i.e., fd.out): running an R-factor calculation from an earlier refcalc, and picking a smaller range than the one of the refcalc. The only way users can ask for this is giving an explicit THEO_ENERGIES.

TODO @amimre: perhaps it would be good to add a note about this in the documentation for THEO_ENERGIES.
TODO @amimre: could you please take care of the test with running a refcalc and then, separately, R on a narrower range?