fix: wrong IV shift used
In new rfactor. The 'currently selected' iv_shift should be the one used for the bounds. This means, especially, no shifts for error calculations (was not the case till now). Also, it is safer, and better, now to use round instead of int, as iv_shift is guaranteed to do the right thing