fix, tests: re-parameterize for new values
I had not realized that many of the tests were actually not taking place. That's because @parametrize eagerly evaluates at class creation. This means that adding NEW KEYS to the class attributes would no create new tests in the subclasses.

This also allowed me to spot some incorrect test values.