fix: occ ignored due to float precision error
When looping searches which cover occupation, a warning is frequently displayed at the beginning of a later delta calculation that occupations are unphysical and will be ignored. It seems that if the occupation range includes 1.0 (100%), and occ_offset = 0.0, that 1.0 + 0.0 = 1.0000000000000002 causes this warning to trigger.

I added some tolerance to avoid this. The alternative would be rounding the values (they will be rounded later anyhow when translating to TensErLEED input).