feat: Best explicit R from fd opt to history
Currently, RUN=6 will not generate information for history.info about the final R-factor, which later on makes it a bit hard to see whether the optimization did anything. This stores the best R-factor that was actually calculated in one of the reference calculation (NOT the parabola minimum, but probably close to it).

@michele-riva @amimre Please check whether this makes sense to also push in master at this time.