fix: graceful exit on MemoryError
When calculating bad pixels, it may happen that the PC does not have enough RAM to allocate the necessary arrays. We should not crash, but report the problem to the user.

This is a partial solution to the problem. In principle it would be possible to do the calculations by (i) having a better look at whether we can free some memory, and (ii) performing the calculations on smaller quadrants of the arrays, properly considering edges.