feat:better phaseshift averaging
Electron scattering phaseshifts may be produced and read in with jumps by pi. Since they factor into the calculation in a pi-periodic manner, this did not cause problems in the past. However, since we are now interpolating phaseshifts in multiple places, we may end up introducing discontinuities or fake behavior when the values wrap around.

This commit introduces some extra checks that corrects potential jumps in the phaseshifts and makes them continuous.

It also introduces an additional check that calculates the variance of phaseshifts for atoms that are designated as equivalent and warns if they are too different. This may help catch averaging over non-equivalent sites, such as different coordination environments.

These ideas came from a discussion with @michele-riva, @schmid-iap and @FortranFlo.