fix: queued connections - end energy bug
Continuous measurements did not set the beam energy to zero at the end of the measurement. This was because the signals were processed in a manner that allowed the measurement to send the change mode command before the ok of the stop command arrived, which led the measurement to disconnect signals before the end energy was actually set. The signals are now queued which means the queued commands of the primary controller will not be executed immediately, but they will be executed in the order they arrived at.