feat: debug message rework
Debug messages used to be sent as a single message in which PC_DEBUG would be the first byte. In order not to confuse these debug messages with regular data messages we decided to split the PC_DEBUG from the debug message. It is sent as a 1 byte message like the PC_OK or PC_ERROR messages. Afterwards the actual message content follows. Debug messages can now be sent during a regular measurement without a problem.