fix: bulk slab for slab with vacuum below
If a slab has some vacuum gap at the bottom, the bulk is incorrectly detected. The reason is the way we detect bulk:
1. Create a thick, fake bulk containing all atoms below the cutoff. This slab also maintains the vacuum gap underneath.
2. reduce the c vector of this thick bulk so as it is as thick as the distance between the bottommost bulk atom and the bottommost non-bulk atom.
3. reduce the unit cell vectors to get the smallest repeat unit.

The problem arises at point 2: the presence of a vacuum gap may cause the topmost bulk atoms to be folded back to the bottom of the thick bulk. This then makes point 3. fail.

The solution is to make sure that the bottommost atom is always at c=0 in the thick bulk.