fix, tests: minimal bulk ab
The previous implementation of the test would use the very same method to gather info about the minimal cell as the one used for the actual implementation of ensure_minimal_bulk_ab_cell.

This means that the test was tautological, and not actually testing for the correct outcome.

Also, we don't really care if the cell that we get out is a rotated version of the one we expected.