fix: bulk slab for slab with vacuum below
This is a second, advanced version of the fix in 1b6b6549d7b3d9a2af47c00deaed5bfd8859fe94. It is better not to force the bottom-most atom to be exactly at c==0, as there can be floating-point rounding errors that may cause the bottommost atom to be re-collapsed to the top of the slab, causing similar issues as with having the vacuum gap at the bottom.