fix: complain if vacuum at bottom
When trying to minimize the c vector for a bulk slab. We could .remove_vacuum_at_bottom, but we would also need to take proper care of re-shifting things back upwards to maintain the centering that is default in the call to appy_bulk_cell_reduction. It seems more prudent to rather have users take care of the shifting themselves. With the current code we should never run into the exception.