fix: z-periodic collapse of BulkSlab coords
When we reduce the unit cell along c, we should make sure to collapse the coordinates in a z-periodic manner (if users ask to). This is important to prevent atoms somewhat close to 1 and somewhat close to 0 that should belong to the same sublayer to end up in different ones.

This partly fixes a bug reported by @fkraushofer for the bulk detection of TiO2. More fixes for that case coming up soon...