fix, tests: incorrect check for centering
Previous implementation would use .atlist[0] after sorting, which, however, was the bottom-most atom, not the topmost one.

Checking for the topmost one is tough, as reducing the c vector may delete the top atoms upon duplicate removal, so one cannot really check that the atom at the top is the same one.