feat: detect and complain if duplicate atoms
For now a method of SurfaceSlab that considers 2D replicas. There does not seem to be a reasonable use case for now for BulkSlab specifically. If this need arises, the method should be re-thought to include a somewhat cyclic check. (Probably it is enough to have the 'None' layer be self.layers[0] if self.is_bulk and self.n_layers >1, and do layer_dist =% c[2]).