fix: check_atom_collision in fd-opt
Closes #103

Together with the implementation of the slab method in 6bd5edc1577dd27982c84c278e71ebf4bb1f36fb and the check at the beginning of the initialization in the same commit.

After a discussion with @fkraushofer, we opted not to check after a search, as the TensErLEED will anyway fail miserably in this case --- and we raise SearchMaxIntensitiesError before even applying the displacements to the slab.