feat: firmware upgrade dialog
This first version of the firmware upgrade dialog can detect Arduino Micro boards and upload firmware from a selected folder to them. Device detection and firmware upload are done in a separate thread as not to stall the GUI. Firmware is taken from a .zip archive and uploaded to the device via the Arduino CLI. The Arduino core and CLI update features have yet to be implemented in the dialog. Box ID detection and installed firmware detection have yet to be implemented. A ViPErLEED error class has yet to be added to enable proper error reporting.