feat, fix: list_devices reworked
list_devices now returns a list containing tuples. (And therefore get_devices returns a dict containing tuples.)  The first element of a tuple is the device name, the second element is a dict that may contain any additional information. This dict will be used forward firmware version information to the firmware upgrade dialog.

fixes #160
Hardware is now always checked and not only if any of the ports has changed. The stalling part of the issue is mentioned in #115 and still needs to be addressed.