feat: SurfaceSlab.check_vacuum_gap
This is a replacement of the less versatile (and, as a matter of fact, buggy) ensure_away_from_c_edges that used to be part of files.poscar. Most of the cases handled now were not handled correctly before.

The new behavior is:
  1. Slab has enough (>5A) vacuum at (or very close to) the top of the slab --> OK
  2. Slab has very very little vacuum (fractional c distance between top and bottom <2e-4), or has a large-enough vacuum gap somewhere in the middle --> raise and stop
  3. Slab has too little vacuum (<5A) --> warn and proceed

In case 1, when the vacuum is very close to, but not at, the top --> atoms are shifted a tad (<2e-4 c fraction), log INFO
In cases 2&3, a new POSCAR_vacuum_corrected is written to SUPP. Users can use it for the next run.

All of the above checks happen on a slab that has  its coordinates collapsed to the base cell.

It is a conscious choice to make this a SurfaceSlab method only. BulkSlab are considered 3D periodic and we don't care about vacuum gaps.

A case that is not covered, and that we cannot reasonably cover: atoms originally outside the cell fold back inside but end up to be lower/higher than the top/bottom atom. This may be a legitimate case or not. The only way to treat this case would be to log some (non-stopping) message if the unit cell to which atoms belong (before collapsing) is not the same for all of them.

As discussed, I'd ask @fkraushofer to take a close look to spot if there's anything incorrect. I'd also like an opinion about the rp.halt level at intialization.py L805, and the option above about atoms 'not in the same cell'.

Of course, @amimre's opinion is also welcome.