fix: create_layers with atoms at c==0
The current implementation of create_layers with auto-cutting uses a strict '<' to select atoms between the cut positions. This is problematic on its own, but even more so considering that our collapsed coordinates are between -COLLAPSE_EPS (included) and 1-COLLAPSE_EPS. This means that any atom between c==-COLLAPSE_EPS and c==0 was never taken into consideration when deciding the positions of cut planes. This was partly obscured by the fact that we were shifting atoms up in poscar.read