fix: do not silently swallow TypeError
helpers.remove_duplicates wuold silentrly swallow all TypeErrors raised during the call to dict.fromkeys. However, it is legitimate for the TypeError not to represent an attempt to compute a hash from a non-hashable object. This is, e.g., the case for PlaneGroup.set_screws_and_glides when a glide-plane-defining string is given without a cell_shape.