fix: Woods negative rotations
After a discussion with @amimre, we opted to always display the full angle corresponding to the matrix given. Angles are positive counter-clockwise.

This has the disadvantage of having multiple strings for reconstructions that are effectively identical. However, this uniquely identifies the matrix (assuming it maintains the handedness of the basis).