fix: Woods stores also matrix version
This is important for surface structures that can be represented by a Wood's notation, but for which the single-digit precision of the rotation angle stored in the string representation is insufficient.

Notice that the to_matrix method does not attempt to fix an angle mismatch, while the setter of .matrix does.

Fixes #179