fix: consistency of Woods string and basis
So far, we were not complaining for weird input, e.g., a rt3xrt3 on an oblique cell.

Checking for this kind of inconsistency made the string setter more complex, and gave a possible maximum-recursion error. That's the reason for almost NEVER using the setter internally, except for the initialization.