fix:indexing by site.label rather than site object
This fixes a hidden left-over bug introduced by the slab-refactor.
In psgen.py, a "subatlist", a dict used to sort and keep track of atoms and sites, used to be indexed by site object and element (str). However, for supercells created when mixing elements, the supercell slab is created as a copy, changing the site-object hash which resulted in KeyErrors when acessing the dict.
I changed the dict key to be the equivalent & unique site.label (e.g. A_surf) which does not change when the site/slab object is copied.

@michele-riva