fix: consistent sorting of beams
Fixes #184.

The problem was the relative sorting of equivalent-beam groups that happen to appear at the same energy. These would be random-sorted -- based on the __hash__ value, which is unstable among different interpreter sessions. The solution is to use an extra criterion involving h&k to resolve sorting for same-energy beams. The same criterion used to sort beams within one equivalence group is used.

@fkraushofer: I'll port this to installable soon. Would you mind checking that this works now? You may need to run one refcalc first to get the beamlist in the stable order. After that, it should never change again.