fix: clear sublayers when making bulkslab
This missing line was the root cause of the last open point of #187. We now use the 'smallest' container available to pick which atom is at the top/bottom of a slab, in the following order: top/bottom sublayer; top/bottom layer; whole slab.

Not clearing the sublayers here would cause non-bulk atoms from slab to be identified as being at the top later in apply_bulk_cell_reduction. This would screw up the centering of the bulk slab.