refactor: tensorleed directory detection
Let run_calc take care of trying to find a suitable tensor-LEED directory. This way we don't have to bother doing it in several other places.

The behavior is also a bit more forgiving. The tensor-LEED code may also be (close to) the current working directory. The precedence is as follows:
1. path given as the source kwarg
2. environment variable VIPERLEED_TENSORLEED
3. current directory

I'd like some feedback on my comment at L288 from @fkraushofer and @amimre: I'm not sure relying on folder names is ideal since we are extracting the tensor-LEED code from the viperleed tree.