fix: remove _try_continuing timer
Removed _try_continuing timer and replaced it's functionality with a call to _ready_for_next_measurement in _on_controller_data_ready. This is to prevent a race condition in which a measurement is over and trying to finalize after getting the last image and the timer fires again.

Seems to fix the bug with trying to get frame rate after a measurement is over. This bug might have been exactly because of this race condition.