fix: doc dependencies in pyproject
The older versions of sphinx have some bug giving "missing number treated as zero" when building latexpdf.

Notice that the version requirements essentially limit the python version for building docs to at least 3.9, as the myst-XXX dependencies are not available in the required versions in 3.8. This is probably acceptable given the fact that likely no-one will build the docs other than the developers.