fix: crash in rfactor with custom IVBEAMS
A bug pointed out by @fkraushofer.
When calculating with a set energy range, but with a non-autogenerated IVBEAMS, the rfactor calculation may crash.

The parameter MNGP in the rfactor-PARAM which determines the size of the array holding the intensity data was set according to the energy range covered by the beams in IVBEAMS. However, the stored reference calculation would still contain the entire energy range, such that this array would be too small.
As a result, out of bounds elements were accessed in the rfactor calculation.
To fix, I increased the MNGP parameter to use the actual size needed.