doc,fix: review plane groups
Fix: avoid vertical overflow of figure in LaTeX.

The current solution is not ideal, as the figure in the PDF is pretty tiny. A second version of the figure (_two_columns) is included in this commit. Unfortunately it cannot yet be used to selectively pick the file depending on the build context because of https://github.com/sphinx-doc/sphinx/issues/4242.

This RST is prepared to handle it (exception: adjust figure size in PDF) as soon as the issue above is resolved.