doc: rename
The :glob: option of toctree sorts the entries by file name. Having a serious mismatch between the parameter name and the file name moves the parameter in an unrelated spot of the toctree.

Take the occasion to also rename a bunch of reference labels that were significantly mismatching with respect to the current names.