fix,doc: better appendices in latex
See comment in index.rst!

The current solution generates an unexpected warning when building HTML:  "WARNING: toctree contains reference to document 'content/appendix' that doesn't have a title: no link will be generated"

This seems to be a bug with the ..only:: directive, as L129 should not be included at all in html. Probably similar to other funny behavior of the same directive.