fix,doc: install intel on windows
The previous documentation would suggest to place a call to the .bat scripts into file pointed to by the AutoRun registry key. Unfortunately, this is NOT POSSIBLE with the current Intel batch files. In fact, they use FOR /F statements for parsing command-line outputs of a DIR call. Since this DIR call is executed in a CMD subprocess, having the .bat script in AutoRun causes INFINTE RECURSION!

This fix proposes a somewhat more tedious, but safe, alternative.