fix: move development dependencies
Moved all development dependencies from pyproject.toml to requirements files. This is necessary as pypi does not allow to have URL-based dependencies in distribution packages, and we use my fork of argparse for the docs.

I think this is also nicer:
- end users don't need to see the extra stuff we need for our development. Also, tests and doc sources are NOT DISTRIBUTED with the sdist, so there is no reason to have dependencies for them.
- installation of the extra dependencies is anyway simple, with pip install [--upgrade] -r requirements-<selection>.txt (and allows to explicitly --upgrade, since we PURPOSELY don't freeze versions)

Notes on using the requirements-[xxx].txt files are now included in the documentation.