fix: pyproject - remove author emails
Seems that having multiple authors with emails is not supported correctly on PyPI: only the first author is rendered, and it may even have the wrong email address. See https://github.com/pypi/warehouse/issues/9400

Remove it also from the maintainers just to be sure not to get the wrong assignment.

Better have fewer metadata that wrong ones...