fix: change consistency check l_max_cutoff to LMAX=3
This cutoff decides the lowest LMAX value used for conssitency checking element ordering in the PHASESHIFT file. As shown in a issue reported by @LutzHammer, the current cutoff of LMAX=4 is slightly too high when using Hydrogen (the lightest scatterer possible). Decreasing to LMAX=3 should allow the function to work correctly.