feat: collect info before running
Changes quite significantly the internal behavior. Earlier we would repeat filesystem calls in multiple spots, especially in property getters. This is not what properties are meant for. Now the relevant information is collected once right before running (or upon user request when update_from_cwd is called) and kept internally in a set of state attributes. The state attributes are cleared as soon as the run finishes to prevent outdated information from being available after a run. This forces users to call update_from_cwd if they're interested in accessing some information about the current tree.

This also removes the annoying side effect of creating history directory, as well as hitory.info and bookkeeper.log files at __init__ of a Bookkeeper instance. They are only created when update_from_cwd is called (which happens automatically during run).