refactor,remove: entry removable
Split off the logics that checks for removability of an entry, as it became quite a big piece of code. Move it to HistoryInfoFile. This also has the advantage that we can remove the very specific .last_entry_has_notes property of a HistoryInfoFile in favor of using directly last_entry.has_notes.