feat: more exception monkeypatches
Generalize the raises_test_exception context manager to a base context manager that raises and checks a given exception. The latter is further implemented as a composite of pytest.raises with a third context manager that just patches a callable to make it raise a given exception.

This allows to simplify quite a bit the tests.