feat,tests: some tests for workhistory
These are partial for now.

Minor refactoring of the bookkeeper after I investigated a bit the structure of workhistory folder names.

I realized that we're doing stuff probably wrongly in bookkeeper DISCARD_FULL for what concerns history folders pulled in from workhistory. This will require a bit of investigation before I can come to a conclusion, and should be followed by an internal discussion.