feat: ExecutionTimer
A timer class that keeps track of when it was started, and is capable of returning the time interval since then. Useful to refactor the code and avoid keeping track of start times as local variables and pass them around.

Important change: replaced the Rparams.starttime with a .timer attribute. The start time is still available with rp.timer.started_at, but that was ever only used to output the 'how long did this take' message.