feat: DateTimeFormat
Date-time formats that we use in multiple spots. This way we avoid code repetition, and we're sure we always use the same format.

Also identified a minor issue: the log entry for calc (used to be tleedm) said the time printed was UTC, while in fact it always been the local time (since the very first commit).