feat: replace dataclass values, include non-init
The default implementation of dataclasses.replace skips completely the copy-over of "special" fields. This include ClassVar (which is good), and init=False fields. The latter is not great not to copy, as it assumes that all the non-init fields are processed during __init__ or __post_init__. This may not be the case for some dataclasses.