feat,refactor: itertools_utils module
Collects iterator stuff. For now, it contains a new version of itertool's pairwise, implemented in terms of a more general n-wise iterator that should be more efficient than the one currently in base, as it does not .tee the iterable.

I'm leaving for now the version in base, pending testing of the current version.