feat: interface for recording states
Implement abstract interface for recording states in the base class. Subclasses can provide a concrete implementation for adding/reverting states, and may override the base-class record and revert methods with the purpose of providing a more explicit signature. Currently this is done for the CalcStateRecorder. However, reverting a state is not yet implemented and raises accordingly.