refactor: don't use splitSublists
Do as suggested by my comment in 212502cc479e71936f9c7d454de2d7bbfe25734d where implementation is straightforward. Currently splitSublists is still used in vibrocc (in combination with the other very funky function readToExc) and in displacements. In both cases, an implementation similar to the one this commit is possible using the portion of the line on the right-hand side of the equals. However, I would reserve this refactoring for different PRs dealing with more substantial changes of the vibrocc and displacements modules (which are very hard to read).