fix: FieldList.remove_fields duplicates
The previous implementation would very likely remove the WRONG fields or FAIL with an IndexError as the indices were not up to date any longer in case of no-duplicates (i.e., there was a valid value-to-index map) as soon as the first item was removed. We pay the price of actually remaking the maps for each removed item rather than having to keep track of possible repeated indices.