fix: FieldList.insert_sorted
The previous implementation was buggy as it would only consider "adjacent" tags to decide where to place a field whose type was not yet in FieldList. It also was unable to properly handle fields with identical duplicate values when looking up the insertion index.

The current implementation fixes the above and also improves the positioning of inserted items with types that are not yet available by using both the previous and the next "bucket" (i.e., tag) to decide the new index. It is chosen so that the new items are closest (in tag-sorting order) to items that already exist. Finally, unknown lines are  always skipped, and items are inserted after them rather than before.