fix: FieldList iteration when making maps
The previous implementation was BUGGED as it would use "for field in self" to iterate through the fields while maps were being updated. The problem was, however, that MutableSequence implements __iter__ via __getitem__, and our __getitem__ tries to look up item indices in the stored maps. This would then return the WRONG items, as the stored maps were likely outdated by the time we would _make_maps.