fix: prevent accumulating white space
Especially true for fields that have a somewhat funny format and that cannot be internally converted to an appropriate data structure. An example: "#TENSORS   1, 2 3". This misses a comma, and is stored as a string. The earlier implementation gobbled up the white spaces into match_['field_value'], and would store "   1, 2 3" as the string value, which could become "   ...   1, 2 3" upon repeated writing/reading to/from string.