feat: support DISCARDED with other text
However, this marks the NotesField as "to be fixed" and NOT discarded. This is because it is hard to judge whether the users just wrote "I'm doing this instead of the previous DISCARDED run" in the notes, or added "because something did not work" to "DISCARDED" on the same line, as in "DISCARDED because something did not work". The former case does not really sound like a DISCARDED entry. At the same time, it is hard to pick the latter as "whenever a line starts with DISCARDED assume it's indeed discarded". A reasonable example would be the user wrapping lines as in "I'm doing this instead of the previous\nDISCARDED run because something went wrong".

We should probably consider to make _DISCARDED a little more "complex", e.g., "!DISCARDED!" or similar, so that people would not likely use the same exact characters when typing they're own notes.