fix,feat: test helpers, dotted name was wrong
make_obj_raise now finds out if the object given is a class or an instance of a class and uses the right name (before, when given a class, it would print "type" instead of the class name).

Also, when given an attribute name, use a dot to separate it from the object name. The previous implementation would show "<obj_name><dotted_name>" instead of "<obj_name>.<dotted_name>".