feat: remove confusing bookkeeper CLI args
Get rid of the --job-name, --work-history-name and --history-name CLI arguments, as per discussion in #198: They are more confusing than useful, and can easily be messed up by users. This also simplifies quite a bit the code of Bookeeper, as we can remove a bunch of init arguments and attributes. The HistoryInfoFile, HistoryInfoEntry, and all their contents still maintain support of job_name for backwards-compatibility reasons.