refactor: use mode to decide behavior
It is more convenient to store the mode in which bookkeeper is currently running and deciding based on that whether (i) the entry that should be added should be marked as DISCARDED (only DISCARD mode), and (ii) whether '_ori'-suffixed files should be used as a fallback when collecting inputs if they're not found in SUPP/original_inputs.

This simplifies a lot by avoiding to pass around arguments to the various methods.

Also, a minor fix: when we pre-archive during DISCARD, there is no reason to try marking the last history.info entry as DISCARDED, as we have already done so. This especially prevents the emission of a "Entry already discarded" log message that would otherwise be confusing.