fix: drop "_ori" when used as fallback
The expected behavior is to collect input files preferentially from SUPP/original_inputs, and falling back to taking those suffixed with "_ori" if the files are missing from original_inputs. However, the current implementation would preserve the "_ori" suffix for those fallback files. This seems unnecessary, as it opens then up problems on (in the future) having to decide which files to pull. The current behavior concerning non-_ori-suffixed files is maintained (i.e., names are suffixed with "_from_root"). I'd like a confirmation from @fkraushofer and @amimre that this is sensible.