feat: bookkeeper metadata file
For now only contains hashes of the history folder in which they will be stored and a reference (via hash) to the "main" folder they were created with (i.e., the one constructed from the input/output files in root). The latter will be used for folders moved to history from workhistory.