feat: don't create history.info at __init__
Rather, remember whether the user wants to create a file if it was missing, and raise where appropriate if it is not.

Also, the path that is now given at __init__ is not the one to the file, but the one to the directory containing it. This allows us to pass along the file name to often.