feat,fix: use HistoryExplorer
Notice also the few changes concerning "adding one" to the job numbers: since we now register both the main as well as the workhistory folders, we don't need to add 1 to get the values right. This also fixes a potential issue in which the "JOB IDS" in history.info would be inconsistent for the workhistory folders moved during a run.

Also fix the tests, as there were folders more recent than the most recent tensors, which makes no sense. (This can only happen for the very first initialization-only run, that is marked as 't000....')